移动和胜利判断写完了

+++ src/main/java/com/edu/xmum/cst206/Interferce/Moveable.java
@@ -1,5 +1,5 @@
 package com.edu.xmum.cst206.Interferce;
 
 public interface Moveable {
-    void move(double dx, double dy);
+    void move(int dx, int dy);
 }
