+++ src/main/java/com/edu/xmum/cst206/Model/GameObject.java
@@ -1,18 +1,13 @@
 package com.edu.xmum.cst206.Model;
 
-import com.edu.xmum.cst206.Interferce.Drawable;
-import javafx.scene.Node;
-import javafx.scene.paint.Color;
-import javafx.scene.shape.Shape;
-
 
 public abstract class GameObject {
-    protected double x;//横坐标
-    protected double y;//纵坐标
+    protected int x;//横坐标
+    protected int y;//纵坐标
     protected double width;//宽
     protected double height;//高
 
-    public GameObject(double x, double y, double width, double height) {
+    public GameObject(int x, int y, double width, double height) {
         this.x = x;
         this.y = y;
         this.width = width;
@@ -21,19 +16,19 @@ public abstract class GameObject {
     //每个游戏对象的绘制方法各自决定
     //每个游戏对象都应该实现对应的get和set方法，以便于其他层模块进行处理
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
