+++ src/main/java/com/edu/xmum/cst206/Model/Player.java
@@ -1,14 +1,6 @@
 package com.edu.xmum.cst206.Model;
 
-import com.edu.xmum.cst206.Interferce.Drawable;
 import com.edu.xmum.cst206.Interferce.Moveable;
-import javafx.scene.Node;
-import javafx.scene.layout.Pane;
-import javafx.scene.paint.Color;
-import javafx.scene.shape.Circle;
-import javafx.scene.shape.Rectangle;
-
-import static com.edu.xmum.cst206.Model.ConstantConfig.CELL_SIZE;
 
 /*
 玩家对象
@@ -17,7 +9,7 @@ import static com.edu.xmum.cst206.Model.ConstantConfig.CELL_SIZE;
 public class Player extends GameObject implements Moveable{
     private double radius;
 
-    public Player(double x, double y, double radius) {
+    public Player(int x, int y, double radius) {
         super(x, y, radius * 2, radius * 2); // 宽度和高度为直径
         this.radius = radius;
     }
@@ -34,7 +26,7 @@ public class Player extends GameObject implements Moveable{
     更新视图的方法由View层负责
      */
     @Override
-    public void move(double dx, double dy) {
+    public void move(int dx, int dy) {
         x += dx;
         y += dy;
     }
